+++ molo/core/cookiecutter/scaffold/{{cookiecutter.directory}}/{{cookiecutter.app_name}}/urls.py
@@ -4,12 +4,18 @@ from django.conf.urls import patterns, include, url
 from django.conf.urls.static import static
 from django.conf import settings
 from django.contrib import admin
+from django.views.generic.base import TemplateView
 
 from wagtail.wagtailadmin import urls as wagtailadmin_urls
 from wagtail.wagtaildocs import urls as wagtaildocs_urls
 from wagtail.wagtailcore import urls as wagtail_urls
 
 from molo.core.views import upload_file, download_file
+
+# Path to a custom template that will be used by the admin
+# site main index view.
+admin.site.index_template = 'django_admin/index.html'
+admin.autodiscover()
 
 # implement CAS URLs in a production setting
 if settings.ENABLE_SSO:
@@ -31,6 +37,9 @@ urlpatterns += patterns(
     url(r'^django-admin/', include(admin.site.urls)),
     url(r'^admin/', include(wagtailadmin_urls)),
     url(r'^documents/', include(wagtaildocs_urls)),
+    url(r'^robots\.txt$', TemplateView.as_view(
+        template_name='robots.txt', content_type='text/plain')),
+    url(r'^sitemap\.xml$', 'wagtail.contrib.wagtailsitemaps.views.sitemap'),
 
 {% for app_name, regex in cookiecutter.include %}
     url(r'{{regex}}',
@@ -40,6 +49,9 @@ urlpatterns += patterns(
 {% endfor %}
     url(r"^mote/", include("mote.urls", namespace="mote")),
     url(r'', include('molo.core.urls')),
+    url(r'^profiles/', include(
+        'molo.profiles.urls',
+        namespace='molo.profiles', app_name='molo.profiles')),
     url('^', include('django.contrib.auth.urls')),
     url(r'', include(wagtail_urls)),
 )