+++ molo/core/tasks.py
@@ -1,3 +1,5 @@
+import random
+
 from datetime import datetime
 
 from celery import task
@@ -10,10 +12,8 @@ from django.core import management
 
 from molo.core.content_import import api
 from molo.core.models import (
-    ArticlePage, Main, SectionIndexPage, SiteLanguage, SectionPage)
+    ArticlePage, Main, SectionIndexPage, SectionPage, Languages, SiteSettings)
 
-from wagtail.contrib.settings.context_processors import SettingsProxy
-from wagtail.wagtailcore.models import Site
 from django.utils import timezone
 
 IMPORT_EMAIL_TEMPLATE = "content_import/import_email.html"
@@ -25,26 +25,34 @@ def rotate_content(day=None):
     """ this method gets the parameters that are needed for rotate_latest
     and rotate_featured_in_homepage methods, and calls them both"""
     # getting the content rotation settings from site settings
-    main_lang = SiteLanguage.objects.filter(is_main_language=True).first()
-    main = Main.objects.all().first()
-    index = SectionIndexPage.objects.live().first()
-    site = Site.objects.get(is_default_site=True)
-    settings = SettingsProxy(site)
-    site_settings = settings['core']['SiteSettings']
-    if day is None:
-        day = datetime.today().weekday()
-    # creates a days of the week list
 
-    # calls the two rotate methods with the necessary params
-    if main and index:
-        rotate_latest(main_lang, index, main, site_settings, day)
-        rotate_featured_in_homepage(main_lang, day)
+    for main in Main.objects.all():
+        site = main.sites_rooted_here.all().first()
+        main_lang = Languages.for_site(site).languages.filter(
+            is_main_language=True).first()
+        index = SectionIndexPage.objects.live().child_of(main).first()
+        site_settings = SiteSettings.for_site(site)
+        if day is None:
+            day = datetime.today().weekday()
+
+        # calls the two rotate methods with the necessary params
+        if main and index:
+            rotate_latest(main_lang, index, main, site_settings, day)
+            rotate_featured_in_homepage(main_lang, day, main)
 
 
 @task(ignore_result=True)
 def publish_scheduled_pages():
     management.call_command(
         'publish_scheduled_pages', verbosity=0, interactive=False)
+
+
+@task(ignore_result=True)
+def clearsessions():
+    # Expired sessions will only be cleared roughly once an hour - randomly
+    if random.randint(0, 59) == 0:
+        management.call_command(
+            'clearsessions', verbosity=0, interactive=False)
 
 
 @task(ignore_result=True)
@@ -104,10 +112,12 @@ def rotate_latest(main_lang, index, main, site_settings, day):
 
     def demote_last_featured_article():
         # set the last featured_in_latest article to false
-        article = main.latest_articles().live().last()
-        article.featured_in_latest_start_date = None
-        article.featured_in_latest_end_date = None
-        article.save_revision().publish()
+        if main.latest_articles().live().count() >= 2:
+            article = main.latest_articles().live().last()
+            article.featured_in_latest = False
+            article.featured_in_latest_start_date = None
+            article.featured_in_latest_end_date = None
+            article.save_revision().publish()
 
     days = get_days_site_settings(site_settings)
     # checks if the current date is within the content rotation range
@@ -125,8 +135,8 @@ def rotate_latest(main_lang, index, main, site_settings, day):
                         random_article = ArticlePage.objects.live().filter(
                             featured_in_latest=False,
                             languages__language__id=main_lang.id
-                        ).descendant_of(index).order_by('?').first()
-
+                        ).descendant_of(index).order_by('?').exact_type(
+                            ArticlePage).first()
                         # set random article to feature in latest
                         if random_article:
                             random_article.featured_in_latest_start_date = \
@@ -136,18 +146,21 @@ def rotate_latest(main_lang, index, main, site_settings, day):
                             demote_last_featured_article()
 
 
-def rotate_featured_in_homepage(main_lang, day):
-    def demote_last_featured_article():
-            article = ArticlePage.objects.live().filter(
+def rotate_featured_in_homepage(main_lang, day, main):
+    def demote_last_featured_article_in_homepage():
+            articles = ArticlePage.objects.descendant_of(main).live().filter(
                 featured_in_homepage=True,
                 languages__language__id=main_lang.id
             ).order_by(
-                '-featured_in_homepage_start_date').last()
-            article.featured_in_homepage_start_date = None
-            article.featured_in_homepage_end_date = None
-            article.save_revision().publish()
+                '-featured_in_homepage_start_date')
+            if articles.count() >= 2:
+                article = articles.last()
+                article.featured_in_homepage = False
+                article.featured_in_homepage_start_date = None
+                article.featured_in_homepage_end_date = None
+                article.save_revision().publish()
 
-    for section in SectionPage.objects.all():
+    for section in SectionPage.objects.descendant_of(main):
         days = get_days_section(section)
         # checks if current date is within the rotation date range
         if section.content_rotation_start_date and \
@@ -163,7 +176,8 @@ def rotate_featured_in_homepage(main_lang, day):
                             random_article = ArticlePage.objects.live().filter(
                                 featured_in_homepage=False,
                                 languages__language__id=main_lang.id
-                            ).child_of(section).order_by('?').first()
+                            ).descendant_of(section).order_by('?').exact_type(
+                                ArticlePage).first()
 
                             # promotes an article and bumps last one off list
                             if random_article:
@@ -172,7 +186,7 @@ def rotate_featured_in_homepage(main_lang, day):
                                     datetime.now()
                                 random_article.save_revision().publish()
                                 promote_articles()
-                                demote_last_featured_article()
+                                demote_last_featured_article_in_homepage()
 
 
 def send_import_email(to_email, context):